<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Redm_package.R
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
>>>>>>> master
\docType{data}
\name{block_3sp}
\alias{block_3sp}
\title{Time series for a three-species coupled model.}
\description{
<<<<<<< HEAD
Time series generated from a discrete-time coupled Lotka-
Volterra model exhibiting chaotic dynamics.
=======
Time series generated from a discrete-time coupled Lotka-Volterra 
model exhibiting chaotic dynamics.
>>>>>>> master
}
\author{
Hao Ye
}

