% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\name{BlockLNLP}
\alias{BlockLNLP}
\title{C++ compiled object for multivariate forecasting.}
\description{
C++ compiled object for multivariate forecasting.
}
\seealso{
\code{\link{block_lnlp}} for the R wrapper.
}

