\docType{data}
\name{circle}
\alias{circle}
\title{2-D timeseries of a circle.}
\format{A data frame with 200 rows and 3 columns: 
\describe{
  \item{\code{Time}}{time index.}
  \item{\code{x}}{sin component.}
  \item{\code{y}}{cos component.}
}}
\usage{
circle
}
\description{
Time series of of circle in 2-D (sin and cos).
}
\keyword{datasets}
