\name{block_gp}
\alias{block_gp}
\alias{ccm_means}
\alias{tde_gp}
\alias{test_nonlinearity}
\title{Deprecated functions}
\usage{
block_gp(block,lib = c(1, NROW(block)),pred = lib,tp = 1,phi = 0,
v_e = 0,eta = 0,fit_params = TRUE,columns = NULL,target_column = 1,
stats_only = TRUE,save_covariance_matrix = FALSE,
first_column_time = FALSE,silent = FALSE, ...)
ccm_means(df, FUN = mean, ...)
tde_gp(time_series,lib = c(1, NROW(time_series)),pred = lib,
E = 1:10,tau = 1,tp = 1,phi = 0,v_e = 0,eta = 0,fit_params = TRUE,
stats_only = TRUE,save_covariance_matrix = FALSE,silent = FALSE, ...)
test_nonlinearity(ts,method = "ebisuzaki", num_surr = 200,
T_period = 1,E = 1, ...)
}
\arguments{
  \item{block}{not implemented}
  \item{lib}{not implemented}
  \item{pred}{not implemented}
  \item{tp}{not implemented}
  \item{phi}{not implemented}
  \item{v_e}{not implemented}
  \item{eta}{not implemented}
  \item{fit_params}{not implemented}
  \item{columns}{not implemented}
  \item{target_column}{not implemented}
  \item{stats_only}{not implemented}
  \item{save_covariance_matrix}{not implemented}
  \item{first_column_time}{not implemented}
  \item{silent}{not implemented}
  \item{df}{not implemented}
  \item{FUN}{not implemented}
  \item{time_series}{not implemented}
  \item{E}{not implemented}
  \item{tau}{not implemented}
  \item{ts}{not implemented}
  \item{method}{not implemented}
  \item{num_surr}{not implemented}
  \item{T_period}{not implemented}
  \item{...}{not implemented}
}
\value{
Not implemneted.
}
\description{
Deprecated functions. 
}
%\examples{
%}
