% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{e054_succession}
\alias{e054_succession}
\title{Succession data at the Cedar Creek LTER}
\description{
Experiment 054 is a subset of the long-term observational study 
of old field succession at the Cedar Creek LTER.
}
\author{
****
}

