% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{e120_biodiversity}
\alias{e120_biodiversity}
\title{Biodiversity data at the Cedar Creek LTER}
\description{
Experiment 120, the "Big Biodiversity" experiment at Cedar Creek
LTER. This experiment is the longest running randomized test for the effects
of plant diversity on ecosystem functions.
}
\author{
****
}

