% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{environmental_data}
\alias{environmental_data}
\title{Physical time series for Fraser River sockeye salmon}
\description{
Time series of Fraser River discharge, sea-surface
temperature at 2 British Columbia lighthouses, and the
Pacific Decadal Oscillation to be used as environmental
covariates of Fraser River sockeye salmon.
}
\author{
Hao Ye
}

