\docType{data}
\name{EvergladesFlow}
\alias{EvergladesFlow}
\title{Water flow to NE Everglades}
\format{A data frame with 1379 rows and 2 columns: 
\describe{
  \item{\code{Date}}{Date.}
  \item{\code{S12CD_S333_CFS}}{Cumulative weekly flow (CFS).}
}
}
\usage{
EvergladesFlow
}
\description{
  Cumulative weekly water flow into northeast Everglades from water
  control structures S12C, S12D and S333 from 1980 through 2005.
}
\keyword{datasets}
