% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm_fun.R
\name{gen_mkt_weights}
\alias{gen_mkt_weights}
\title{Generate market weights}
\usage{
gen_mkt_weights(crops, markets)
}
\arguments{
\item{crops}{A character vector of crops.}

\item{markets}{A character vector of consumer markets.}
}
\value{
A character vector of weights.
 @export
}
\description{
Generates weights for types of consumers.
}
