% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm_fun.R
\name{gmw}
\alias{gmw}
\title{Generate Market Weights and Put them in a Tibble}
\usage{
gmw(crops, markets)
}
\arguments{
\item{crops}{A character vector of crops.}

\item{markets}{A character vector of consumer markets.}
}
\value{
Returns a tibble of the market weights.
 @export
}
\description{
Uses `gen_mkt_weights` to create a tibble for storing the market weights.
}
