% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\name{LNLP}
\alias{LNLP}
\title{C++ compiled object for univariate forecasting.}
\description{
C++ compiled object for univariate forecasting.
}
\author{
Hao Ye
}
\seealso{
\code{\link{simplex}} and \code{\link{s_map}} for R wrappers.
}

