\docType{data}
\name{Lorenz5D}
\alias{Lorenz5D}
\title{5-D Lorenz'96}
\format{Data frame with 1000 rows and 6 columns
\describe{
  \item{\code{Time}}{Time.}
  \item{\code{V1}}{variable 1.}
  \item{\code{V2}}{variable 2.}
  \item{\code{V3}}{variable 3.}
  \item{\code{V4}}{variable 4.}
  \item{\code{V5}}{variable 5.}
}
}
\usage{
Lorenz5D
}
\description{5-D Lorenz'96 timeseries with F = 8.
}
\references{
Lorenz, Edward (1996). Predictability - A problem partly solved,
Seminar on Predictability, Vol. I, ECMWF.
}
\keyword{datasets}
