\docType{data}
\name{paramecium_didinium}
\alias{paramecium_didinium}
\title{Time series for the Paramecium-Didinium laboratory experiment}
\usage{
paramecium_didinium
}
\description{
Time series of Paramecium and Didinium abundances (#/mL) from 
an experiment by Veilleux (1979)
}
\keyword{datasets}
