% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Redm-package}
\alias{Redm}
\alias{Redm-package}
\title{Applications of empirical dynamic modeling from time series.}
\description{
The Redm package provides an interface from R to C++
compiled objects that use time delay embedding to perform
state-space reconstruction and nonlinear forecasting.
}
\author{
Hao Ye
}
\keyword{package}

