% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{sardine_anchovy_sst}
\alias{sardine_anchovy_sst}
\title{Time series for the California Current Anchovy-Sardine-SST system}
\format{\describe{
  \item{\code{year}}{year of measurement}
  \item{\code{anchovy}}{anchovy landings, scaled to mean = 0, sd = 1}
  \item{\code{sardine}}{sardine landings, scaled to mean = 0, sd = 1}
  \item{\code{sio_sst}}{3-year running average of sea surface temperature at 
    SIO pier, scaled to mean = 0, sd = 1}
  \item{\code{np_sst}}{3-year running average of sea surface temperature at 
    Newport pier, scaled to mean = 0, sd = 1}
}}
\usage{
sardine_anchovy_sst
}
\description{
Time series of Pacific sardine landings (CA), Northern anchovy 
  landings (CA), and sea-surface temperature (3-year average) at the SIO 
  pier and Newport pier
}
\author{
****
}
\keyword{datasets}
