% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sockeye_data}
\alias{sockeye_data}
\title{Biological time series for Fraser River sockeye salmon}
\description{
Time series of recruitment, juveniles, and effective female
spawners for 19 stocks of sockeye salmon in the Fraser
River system.
}
\author{
Hao Ye
}

