\docType{data}
\name{TentMap}
\alias{TentMap}
\title{Time series for a tent map with mu = 2.}
\format{Data frame with 999 rows and 2 columns
\describe{
  \item{\code{Time}}{time index.}
  \item{\code{TentMap}}{tent map values.}
}
}
\usage{
TentMap
}
\description{
First-differenced time series generated from the tent map
  recurrence relation with mu = 2.
}
\keyword{datasets}
