% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Redm_package.R
\docType{data}
\name{tentmap_1d}
\alias{tentmap_1d}
\title{Time series for a tent map with mu = 2.}
\description{
Time series generated from the tent map recurrence relation
with mu = 2.
}
\author{
Hao Ye
}

