% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{tentmap_del}
\alias{tentmap_del}
\title{Time series for a tent map with mu = 2.}
\description{
First-differenced time series generated from the tent map recurrence relation with mu = 2.
}
\author{
Hao Ye
}

