\docType{data}
\name{TentMapNoise}
\alias{TentMapNoise}
\title{Time series of tent map plus noise.}
\format{Data frame with 999 rows and 2 columns
\describe{
  \item{\code{Time}}{time index.}
  \item{\code{TentMap}}{tent map values.}
}
}
\usage{
TentMapNoise
}
\description{
First-differenced time series generated from the tent map
  recurrence relation with mu = 2 and random noise.
}
\keyword{datasets}
