\docType{data}
\name{Thrips}
\alias{Thrips}
\title{Apple-blossom Thrips time series}
\description{
Seasonal outbreaks of Thrips imaginis.
}
\references{
Davidson and Andrewartha, Annual trends in a natural population of
Thrips imaginis \emph{Thysanoptera}, Journal of Animal Ecology, 17,
193-199, 1948.
}
