% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{two_species_model}
\alias{two_species_model}
\title{Time series for a two-species coupled model.}
\format{A data frame with 1000 rows and 3 variables: 
\describe{
  \item{\code{time}}{time index (# of generations)}
  \item{\code{x}}{abundance of simulated species x}
  \item{\code{y}}{abundance of simulated species y}
}}
\usage{
two_species_model
}
\description{
Time series generated from a discrete-time coupled 
  Lotka-Volterra model exhibiting chaotic dynamics.
}
\author{
Hao Ye
}
\keyword{datasets}
