% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\name{Xmap}
\alias{Xmap}
\title{C++ compiled object for convergent cross mapping.}
\description{
C++ compiled object for convergent cross mapping.
}
\seealso{
\code{\link{ccm}} for the R wrapper.
}

