% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{block_3sp}
\alias{block_3sp}
\title{Time series for a three-species coupled model.}
\format{A data frame with 200 rows and 10 variables: 
\describe{
  \item{\code{time}}{time index (# of generations)}
  \item{\code{x_t}}{abundance of simulated species $x$ at time $t$}
  \item{\code{x_t-1}}{abundance of simulated species $x$ at time $t-1$}
  \item{\code{x_t-2}}{abundance of simulated species $x$ at time $t-2$}
  \item{\code{y_t}}{abundance of simulated species $y$ at time $t$}
  \item{\code{y_t-1}}{abundance of simulated species $y$ at time $t-1$}
  \item{\code{y_t-2}}{abundance of simulated species $y$ at time $t-2$}
  \item{\code{z_t}}{abundance of simulated species $z$ at time $t$}
  \item{\code{z_t-1}}{abundance of simulated species $z$ at time $t-1$}
  \item{\code{z_t-2}}{abundance of simulated species $z$ at time $t-2$}
}}
\usage{
block_3sp
}
\description{
Time series generated from a discrete-time coupled 
  Lotka-Volterra model exhibiting chaotic dynamics.
}
\author{
Hao Ye
}
\keyword{datasets}
