% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{block_3sp}
\alias{block_3sp}
\title{Time series for a three-species coupled model.}
\description{
Time series generated from a discrete-time coupled
Lotka-Volterra model exhibiting chaotic dynamics.
}
\author{
Hao Ye
}

