% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccm_interface.R
\name{ccm_means}
\alias{ccm_means}
\title{Take output from ccm and compute means as a function of library size.}
\usage{
ccm_means(ccm_df)
}
\arguments{
  \item{ccm_df}{a data.frame, usually output from the
  \code{\link{ccm}} function}
}
\value{
A data.frame with forecast statistics averaged at each
unique library size
}
\description{
\code{ccm_means} is a utility function to summarize output
from the \code{\link{ccm}} function
}

