% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_block}
\alias{make_block}
\title{Make a lagged block for multiview}
\usage{
make_block(block, t = NULL, max_lag = 3, tau = 1, lib = NULL,
  restrict_to_lib = TRUE)
}
\arguments{
\item{block}{a data.frame or matrix where each column is a time series}

\item{t}{the time index for the block}

\item{max_lag}{the total number of lags to include for each variable. So if 
max_lag == 3, a variable X will appear with lags X[t], X[t - tau], 
X[t - 2*tau]}

\item{tau}{the lag to use for time delay embedding}

\item{lib}{a 2-column matrix (or 2-element vector) where each row specifies 
the first and last *rows* of the time series to use for attractor 
reconstruction}

\item{restrict_to_lib}{whether to restrict the final lagged block to 
just the rows specified in lib (if lib exists)}
}
\value{
A data.frame with the lagged columns and a time column. If the 
  original block had columns X, Y, Z and max_lag = 3, then the returned 
  data.frame will have columns TIME, X, X_1, X_2, Y, Y_1, Y_2, Z, Z_1, Z_2.
}
\description{
\code{\link{make_block}} generates a lagged block with the appropriate max_lag and 
tau, while respecting lib (by inserting NANs, when trying to lag past lib 
regions)
}
\examples{
data("block_3sp")
make_block(block_3sp[, c(2, 5, 8)])
}
