% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{paramecium_didinium}
\alias{paramecium_didinium}
\title{Time series for the Paramecium-Didinium laboratory experiment}
\description{
Time series of Paramecium and Didinium abundances (#/mL)
from an experiment by Veilleux (1979)
}
\author{
Hao Ye
}

