<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Redm_package.R
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
>>>>>>> master
\docType{data}
\name{paramecium_didinium}
\alias{paramecium_didinium}
\title{Time series for the Paramecium-Didinium laboratory experiment}
\description{
Time series of Paramecium and Didinium abundances (#/mL) from 
an experiment by Veilleux (1979)
}
\author{
Veilleux
}

