% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{package}
\name{rEDM-package}
\alias{Redm}
\alias{rEDM-package}
\title{Applications of empirical dynamic modeling from time series.}
\description{
The rEDM package provides an interface from R to C++
compiled objects that use time delay embedding to perform
state-space reconstruction and nonlinear forecasting.
}
\author{
Hao Ye
}
\keyword{package}

