% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{sardine_anchovy_sst}
\alias{sardine_anchovy_sst}
\title{Time series for the California Current Anchovy-Sardine-SST system}
\description{
Time series of Pacific sardine landings (CA), Northern anchovy landings (CA), 
and sea-surface temperature (3-year average) at the SIO pier and Newport pier
}
\author{
****
}
