% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lnlp_interface.R
\name{simplex}
\alias{simplex}
\title{Perform univariate forecasting using simplex projection}
\usage{
simplex(time_series, lib = c(1, NROW(time_series)), pred = lib,
  norm_type = c("L2 norm", "L1 norm"), E = 1:10, tau = 1, tp = 1,
  num_neighbors = "e+1", stats_only = TRUE, exclusion_radius = NULL,
  epsilon = NULL, silent = FALSE)
}
\arguments{
\item{time_series}{either a vector to be used as the time series, or a
data.frame or matrix with at least 2 columns (in which case the first column
will be used as the time index, and the second column as the time series)}

\item{lib}{a 2-column matrix (or 2-element vector) where each row specifes the
first and last *rows* of the time series to use for attractor reconstruction}

\item{pred}{(same format as lib), but specifying the sections of the time
series to forecast.}

\item{norm_type}{the distance function to use. see 'Details'}

\item{E}{the embedding dimensions to use for time delay embedding}

\item{tau}{the lag to use for time delay embedding}

\item{tp}{the prediction horizon (how far ahead to forecast)}

\item{num_neighbors}{the number of nearest neighbors to use (any of "e+1",
"E+1", "e + 1", "E + 1" will peg this parameter to E+1 for each run, any
value < 1 will use all possible neighbors.)}

\item{stats_only}{specify whether to output just the forecast statistics or
the raw predictions for each run}

\item{exclusion_radius}{excludes vectors from the search space of nearest
neighbors if their *time index* is within exclusion_radius (NULL turns
this option off)}

\item{epsilon}{excludes vectors from the search space of nearest neighbors
if their *distance* is farther away than epsilon (NULL turns this option
off)}

\item{silent}{prevents warning messages from being printed to the R console}
}
\value{
If stats_only, then a data.frame with components for the parameters
  and forecast statistics:
\tabular{ll}{
  E \tab embedding dimension\cr
  tau \tab time lag\cr
  tp \tab prediction horizon\cr
  nn \tab number of neighbors\cr
  num_pred \tab number of predictions\cr
  rho \tab correlation coefficient between observations and predictions\cr
  mae \tab mean absolute error\cr
  rmse \tab root mean square error
}
Otherwise, a list where the number of elements is equal to the number of runs
  (unique parameter combinations). Each element is a list with the following
  components:
\tabular{ll}{
  params \tab data.frame of parameters (E, tau, tp, nn)\cr
  model_output \tab data.frame with columns for the time index, observations,
    and predictions\cr
  stats \tab data.frame of forecast statistics (num_pred, rho, mae, rmse)\cr
}
}
\description{
\code{simplex} uses time delay embedding on a single time series to
generate an attractor reconstruction, and then applies the simplex
projection algorithm to make forecasts. This method is typically applied,
and the embedding dimension varied, to find an optimal embedding dimension
for the data.
}
\details{
The default parameters are set so that passing a time series as the only
argument will run over E = 1:10 (embedding dimension), using leave-one-out
cross-validation over the whole time series, and returning just the forecast
statistics.

norm_type "L2 norm" (default) uses the typical Euclidean distance:
\deqn{distance(a,b) := \sqrt{\sum_i{(a_i - b_i)^2}}}{distance(a, b) := \sqrt(\sum(a_i - b_i)^2)}
norm_type "L1 norm" uses the Manhattan distance:
\deqn{distance(a,b) := \sum_i{|a_i - b_i|}}{distance(a, b) := \sum|a_i - b_i|}
}

