% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{sockeye_returns}
\alias{sockeye_returns}
\title{Time series for sockeye salmon returns.}
\description{
Time series of sockeye salmon returns from the Fraser River
in British Columbia, Canada.
}
\author{
Hao Ye
}

