% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{two_species_model}
\alias{two_species_model}
\title{Time series for a two-species coupled model.}
\description{
Time series generated from a discrete-time coupled
Lotka-Volterra model exhibiting chaotic dynamics.
}
\author{
Hao Ye
}

