% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEDM_package.R
\docType{data}
\name{two_species_model}
\alias{two_species_model}
\title{Time series for a two-species coupled model.}
\description{
Time series generated from a discrete-time coupled Lotka-Volterra
model exhibiting chaotic dynamics.
}
\author{
Hao Ye
}

